[Systems] Added ssh tunnel to connection
The rover will open a tunnel online at engineer.blue:11380
11322 and 11311

First check if the tunnel is alive
http://engineer.blue:11380/

ROS!
________________________________________
18:45:49 up 1 min,  1 user,  load average: 4.31, 1.57, 0.56

It should output the uptime of the machine.

Your ROS master is:
export ROS_MASTER_URI=http://engineer.blue:11311/

To SSH into the Robot
ssh earth@engineer.blue -p11322

Our ultra-extra secure password for the occasion is still:
earth123

Internally you can jump into the Jetson which is on an internal network
ssh earth@192.168.0.25

If we need the VNC I can open them later too.

Don’t forget this is a live machine with WHEELS!!! which is sitting in a table :D.
Careful when you move it hahaha.

Signed-off-by: Sergio A. Mart <sergio@earthrover.cc>